# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the Organizer application, providing a way to manage personal productivity modules.",
        "title": "Organizer API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support Team",
            "url": "https://www.example.com/support",
            "email": "kiv.426@google.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Authenticates a user and returns a JWT token upon successful login.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Log in an existing user",
                "parameters": [
                    {
                        "description": "User Login Credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.LoginInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully logged in",
                        "schema": {
                            "$ref": "#/definitions/handlers.UserAuthResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error or invalid input (e.g., {\\\"error\\\": \\\"Invalid input: ...\\\"})",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials (e.g., {\\\"error\\\": \\\"Invalid credentials\\\"})",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error (e.g., {\\\"error\\\": \\\"Database error finding user\\\"})",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Creates a new user account and returns a JWT token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User Registration Data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.RegisterInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully registered",
                        "schema": {
                            "$ref": "#/definitions/handlers.UserAuthResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error or invalid input (e.g., {\\\"error\\\": \\\"Invalid input: Key: 'RegisterInput.Email' Error:Field validation for 'Email' failed on the 'email' tag\\\"})",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "409": {
                        "description": "User with this email already exists (e.g., {\\\"error\\\": \\\"User with this email already exists\\\"})",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error (e.g., {\\\"error\\\": \\\"Failed to hash password\\\"})",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/auth/validate-token": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Validates the provided JWT and returns user information if valid.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Validate JWT Token and get user info",
                "responses": {
                    "200": {
                        "description": "User data for valid token",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "401": {
                        "description": "Unauthorized or invalid token (e.g., {\\\"error\\\": \\\"Unauthorized or invalid token\\\"})",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/knowledge-links": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of knowledge links for the current user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "knowledge-links"
                ],
                "summary": "Get all knowledge links for the authenticated user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.KnowledgeLink"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error (e.g., {\"error\": \"Failed to retrieve knowledge links\"})",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a new knowledge link for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "knowledge-links"
                ],
                "summary": "Create a new knowledge link",
                "parameters": [
                    {
                        "description": "Knowledge Link data",
                        "name": "link",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateKnowledgeLinkInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.KnowledgeLink"
                        }
                    },
                    "400": {
                        "description": "Invalid input (e.g., {\"error\": \"Invalid input: ...\"})",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error (e.g., {\"error\": \"Failed to create knowledge link\"})",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/knowledge-links/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a specific knowledge link by its ID, if it belongs to the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "knowledge-links"
                ],
                "summary": "Delete a knowledge link by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Knowledge Link ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Knowledge link deleted successfully (e.g., {\"message\": \"Knowledge link deleted successfully\"})",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Knowledge link not found or access denied",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error (e.g., {\"error\": \"Failed to delete knowledge link\"})",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateKnowledgeLinkInput": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "handlers.LoginInput": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                }
            }
        },
...
        "handlers.RegisterInput": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "age": {
                    "type": "integer",
                    "example": 30
                },
                "contacts": {
                    "type": "string",
                    "example": "+1234567890"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "fullname": {
                    "type": "string",
                    "example": "John Doe"
                },
                "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "password123"
                }
            }
        },
...
        "handlers.UserAuthResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user": {
                    "description": "User model without PasswordHash",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                }
            }
        },
        "models.KnowledgeLink": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "contacts": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "telegramHash": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token. Example: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}